fix(steamos-update): Get progress bar working again
Universal Blue update isn't verbose unless something is logged or an
exception is raised. Because of this, we can't just monitor its output
and increment the progress bar. Instead, we need to wait for it to
complete.

Since there isn't an entirely clean way of doing that, we'll just wait
for the creation of '/tmp/update-check'

This also removes output from the updater to ensure that formatting
stays consistent when the output is passed to Steam.