fix(decky): Ensure we retrieve the user's home directory
Each newline in a just recipe instantiates a new shell unless we explicitly
define the environment. Similarly, because of this limitation, we need to
retrieve the users home directory (since we are executing this recipe with
sudo) so that it does not assume that the user is root

We will have to do this in the future for other recipes that depend on the
environment being preserved. Further, I ran into a limitation where global
env variables must be strings/characters. We can not reuse these variables
either so they must be declared for each recipe

See: https://just.systems/man/en/chapter_44.html