feat: Preinstall GOverlay (#572)
GOverlay (https://github.com/benjamimgois/goverlay)

Close this obviously if this is out of scope for the project or if I did this PR wrong.  

It's a nice GUI for MangoHud, vkBasalt, and ReplaySorcery though.  In my testing in layering it to Bazzite, it complains MangoHud is missing, but it worked in my testing so I'm not sure why it does that.